refactor(proxy-session): derive documentation routes from Xcode process routes
