+++ Sources/ProxySession/Session/XcodeProcessRoute.swift
@@ -1,0 +1,16 @@
+import Foundation
+import ProxyCore
+
+package struct XcodeProcessRoute: Sendable, Equatable {
+    package let target: XcodeProcessTarget
+    package let upstreamIndices: [Int]
+
+    package var primaryUpstreamIndex: Int? {
+        upstreamIndices.first
+    }
+
+    package init(target: XcodeProcessTarget, upstreamIndices: [Int]) {
+        self.target = target
+        self.upstreamIndices = upstreamIndices
+    }
+}
